# avoid quadratic time processing in solver-(assert,min/maximize)